Annotations work final, plus respStmt
Added the rest of the annotations, save three which are indicated via
notes attached to cells in the GDoc.

Added the respStmt with IDs for all annotating participants (plus
myself just in case), and made references to those IDs from every note.
This can be used in the Notes filter.